# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API on Golang for film library service.",
        "title": "Film library service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/actor/change": {
            "post": {
                "description": "Change actor data",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Actors"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/change_actor_data.ChangeActorDataIn"
                        }
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        },
        "/actor/create": {
            "post": {
                "description": "Creating new actor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actors"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/create_actor.CreateActorIn"
                        }
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/create_actor.CreateActorOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        },
        "/actor/delete": {
            "delete": {
                "description": "Delete actor",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Actors"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delete_actor.DeleteActorIn"
                        }
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        },
...
        "/film/change": {
            "post": {
                "description": "Change film data",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/change_film_data.ChangeFilmIn"
                        }
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        },
        "/film/create": {
            "post": {
                "description": "Insert new film",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/create_film.CreatFilmIn"
                        }
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/create_film.CreateFilmOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        },
        "/film/delete": {
            "delete": {
                "description": "Delete all data about film",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delete_film.DeleteFilmIn"
                        }
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorOut"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "change_actor_data.ChangeActorDataIn": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "1976-05-25"
                },
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "name": {
                    "type": "string",
                    "example": "Cillian Murphy"
                }
            }
        },
        "change_film_data.ChangeFilmIn": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Leonardo Dicaprio"
                    ]
                },
                "description": {
                    "type": "string",
                    "example": "Description"
                },
                "id": {
                    "type": "integer",
                    "example": 3
                },
                "rating": {
                    "type": "integer",
                    "example": 10
                },
                "release_date": {
                    "type": "string",
                    "example": "1998-02-20"
                },
                "title": {
                    "type": "string",
                    "example": "Titanic"
                }
            }
        },
        "common.ErrorOut": {
            "type": "object",
            "properties": {
                "error_message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "create_actor.CreateActorIn": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "1976-05-25"
                },
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "name": {
                    "type": "string",
                    "example": "Cillian Murphy"
                }
            }
        },
        "create_actor.CreateActorOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "create_film.CreatFilmIn": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Leonardo Dicaprio"
                    ]
                },
                "description": {
                    "type": "string",
                    "example": "Description"
                },
                "rating": {
                    "type": "integer",
                    "example": 10
                },
                "release_date": {
                    "type": "string",
                    "example": "1998-02-20"
                },
                "title": {
                    "type": "string",
                    "example": "Titanic"
                }
            }
        },
        "create_film.CreateFilmOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "delete_actor.DeleteActorIn": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "delete_film.DeleteFilmIn": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
    }
}